feat: add support for whole hunk matching
Add a `-x` flag (similar to grep's whole line modifier) to require the
pattern to match the entire content of the hunk (or _not_ to match any
line in the inverse case). This would be intended to be used in
conjunction with one or more of the -d/-i/-a flags to get the desired
behaviour.

Also update `go-diff` import path and add go.mod/go.sum